ASF-4:вынес форму со страницы в отдельный компонент
