+++ src/layouts/guest/GuestLayout.tsx
@@ -1,13 +1,12 @@
-// import { Outlet } from "react-router-dom"
-import { Dialog } from "src/ui/dialog/Dialog"
-import style from "./guestLayout.module.scss"
+import { Outlet } from "react-router-dom"
+
+import style from "./GuestLayout.module.scss"
 
 export const GuestLayout = () => {
     return (
         <div className={style.auth}>
             <div className={style.auth__body}>
-                {/* <Outlet/> */}
-                <Dialog show/>
+                <Outlet/>
             </div>
         </div>
     )
